fix: douyin stream preview stall
Using a separate version of shaka-player to avoid stalls on the special stream,
which is mixed with raw AAC data segments.

shaka-player has made some fixes on this:
https://github.com/shaka-project/shaka-player/pull/6360

However, in recent versions, this bug seems to have reappeared,
so we changed to use an older version of shaka-player for the
douyin stream to avoid this.